sched: use cuml demand for idle selection
Currently we choose a idle cpu if its in a shallower c state than
the previous one. But if it comes across a idle cpu with the same
cstate, it skips it.

Instead of skipping it, check if it has a lower cuml demand and if
so prefer that cpu over the previous one. But prefer previous cpu
over a cpu with lower cuml demand.

Change-Id: I94b74b1b1c91d1cc7d4b5f40fb4f79558271e992
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: spezi77 <spezi7713@gmx.net>