cpu_input_boost: implement multi-level Dynamic SchedTune Boost
Dynamic SchedTune Boost occurs on-the-fly during interactions, such
as touch input. This idea was conceived because of an existing battery
drain issue whereby setting a permanent schedtune boost would cause
boosting to occur even when the phone was idling. For example, if a user
sets the /dev/stune/top-app/schedtune.boost to 10, the device may not
idle at the lowest frequency step, which can lead to higher idle and
active drain rates. Thus it should consume less power to boost only
during interactions.

This implementation of Dynamic SchedTune Boost support separate boost
levels for different cpu_input_boost frequency levels. By default,
max boost is 50, input boost is 35, and general boost is 15. The default
values can be changed as necessary in Kconfig. It is also possible to
tune these values dynamically from userspace without reconfiguring the
kernel via the following tunables in sysfs:
  - /sys/module/cpu_input_boost/parameters/dynamic_stune_boost (input)
  - /sys/module/cpu_input_boost/parameters/max_stune_boost (max)
  - /sys/module/cpu_input_boost/parameters/general_stune_boost (general)

The input_stune_boost parameter is exposed as dynamic_stune_boost for
compatibility purposes as this is the parameter users are most likely
going to tune in kernel manager apps.

Setting a value of 0 for any of the boost levels disables boosting for
that level.

History of the original/base implementation by joshuous in the cpu-boost
driver:
  v1 worked by changing the value returned by schedtune_cpu_boost(), but
  did not change the true boost value of the top-app cgroup. In comparison,
  v2 modifies the value in /dev/stune/top-app/schedtune.boost directly.
  v3 reworks the entire Dynamic SchedTune Boost code and improves detection
  of top-app or any other SchedTune group.

This implementation uses v3 of the Dynamic SchedTune Boost framework.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>