cpu_input_boost: introduce display-on stune boost floor
There is a substantial performance increase when the device is in use
from using a floor of 1 for top-app stune boost. The battery life
penalty is negligible and potentially actually positive because it is
less work for the scheduler to iterate 4 cores (the big cluster) first
which has relatively low load, resulting in less scheduler overhead,
resulting in the work getting done faster overall for the device to
spend more time idling. The little cluster is used much more and has
higher load, so the scheduler needs to do more work and iteration to
find an available core.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>