cpu_input_boost: add separate return frequencies for screen-on states
In some cases, a higher minimum frequency for any of the clusters may be
desired when the display is on and the device is in active use. For
example, on sdm845 CAF and Google use 576 MHz as the minimum frequency
for the little cluster at all times instead of the true minimum, 300
MHz. This should offer a decent increase in performance while having a
negligible impact on battery drain. However, to maximize idle battery
life, it would be good to use this only when the display is on. When the
display turned off, both clusters will be reverted to the standard return
frequencies.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>