defconfig: configure minimum screen-on CPU frequencies
RenderBroken and joshchoo from the RenderZenith EAS kernel for OP6 are
under the impression that this is optimal, at least for active screen-on
cases:
"I was told by a third-party that they measured decent performance
increases and negligible battery drains for the higher default min
freqs"

Use 576 MHz as the minimum frequency for the little cluster because of
the above reasoning since this is what CAF and Google use by default.
For now, the big cluster's minimum can stay at the true minimum at all
times.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>