state_notifier: Clean up display state notifier
This makes state_notifier a completely independent driver that does not
need edits to any other drivers in the kernel. It's also cleaner and
more resililent to changes, with lower chances of breaking in unusual
circumstances as this method of display state notification has been
battle-tested in several drivers and perfected by Google and Qualcomm.
Our previous hooks directly in the panel poweroff/poweron code may have
triggered at the wrong times, as msm_drm_notifier events have a
different code path that also checks for duplication and such.

Additionally, a panel that is powered on is not necessarily displaying a
picture. Powering the panel on is not the same thing as unblanking. The
old hook made this assumption.

An added benefit of the new notification source is that an always-on
ambient display will now be considered display-off as it should be,
which will reduce idle power consumption for AOD users.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>